Updated DynamicHelper Coomponent for a more resuasable one
