+++ Source/Applications/MiMD/wwwroot/Scripts/TSX/MiMD/PRC-002/ChangeOverview/ResolveRecord.tsx
@@ -26,7 +26,7 @@ import * as _ from 'lodash';
 import * as PRC002 from '../ComplianceModels';
 import { LoadingIcon, Modal, Warning } from '@gpa-gemstone/react-interactive';
 import { Input, Select } from '@gpa-gemstone/react-forms';
-import { DynamicHelper } from '../Common/DynamicHelper'
+import { HelperTable } from "../../CommonComponents/HelperTable"
 
 declare let homePath: string;
 
@@ -229,6 +229,26 @@ const ConfigFieldEdit = (props: { Field: PRC002.IConfigField, Setter: (record: P
     const NumberChecks = [{ Value: '=', Label: '=' }, { Value: '<>', Label: '<>' }, { Value: '>', Label: '>' }, { Value: '<', Label: '<' }];
     const TextChecks = [{ Value: '=', Label: '=' }, { Value: '<>', Label: '<>' }, { Value: 'IN', Label: 'In' }];
     const [showFunctionHelp, setShowFunctionHelp] = React.useState(false);
+    const help = [
+        { Name: 'PreVal', Description: 'The previous value of your config file value', Example: 'PreVal' },
+        { Name: '+, -', Description: 'Additive', Example: '100 + 7' },
+        { Name: '*, /, %', Description: 'Multiplicative', Example: '100 * 2 / (3 % 2)' },
+        { Name: '^', Description: 'Power', Example: '2 ^ 16' },
+        { Name: '-', Description: 'Negation', Example: '-6 + 10' },
+        { Name: '+', Description: 'String Concatenation', Example: '"abc" + "def"' },
+        { Name: '<<, >>', Description: 'Shift', Example: '0x80 >> 2' },
+        { Name: '=, <>, <, >, <=, >=', Description: 'Comparison', Example: '2.5 > 100' },
+        { Name: 'And, Or, Xor, Not', Description: 'Logical', Example: '(1 > 10) and (true or not false)' },
+        { Name: 'And, Or, Xor, Not', Description: 'Bitwise', Example: '100 And 44 or (not 255)' },
+        { Name: 'If', Description: 'Conditional', Example: 'If(a > 100, "greater", "less")' },
+        { Name: 'Cast', Description: 'Cast and conversion', Example: 'cast(100.25, int)' },
+        { Name: 'String literal', Description: '', Example: '"string!"' },
+        { Name: 'Char literal', Description: '', Example: '\'c\'' },
+        { Name: 'Boolean literal', Description: '', Example: 'true AND false' },
+        { Name: 'Real literal', Description: 'Double and single', Example: '100.25 + 100.25f' },
+        { Name: 'Integer literal', Description: 'Signed/unsigned 32/64 bit', Example: '100 + 100U + 100L + 100LU' },
+        { Name: 'Hex literal', Description: '', Example: '0xFF + 0xABCDU + 0x80L + 0xC9LU' }
+    ]
 
     function ValidValue(): boolean {
         return (props.Field.Value != null && props.Field.Value.length > 0)
@@ -249,7 +269,7 @@ const ConfigFieldEdit = (props: { Field: PRC002.IConfigField, Setter: (record: P
         <button type="button" className="btn btn-light float-right" onClick={() => setShowFunctionHelp(true)}>
             <i style={{ color: '#007BFF' }} className="fa fa-2x fa-question-circle"></i>
         </button>
-        {showFunctionHelp && <DynamicHelper isOpen={showFunctionHelp} onClose={() => setShowFunctionHelp(false)} />}
+        {showFunctionHelp && <HelperTable Data={help} Title={"Dynamic Expression Examples"} IsOpen={showFunctionHelp} onClose={() => setShowFunctionHelp(!showFunctionHelp)} />}
     </>)
 }
